[OMPL] also check constraints in StateValidityCallback
This allows the IK solver to continue finding a valid solution, which also satisfies the kinematic constraints.

In the old situation the IK solver would return a solution, which it thinks is valid, but it is not. The sampler will keep calling the IK solver. Though for IK solvers, which are not that random. There is a high chance of returning the same solution over and over. Which keeps failing the constraints.

By checking the constraints in the callback the IK solver will continue and return a valid solution.